resolve circ imports
