+++ src/artemis/experiment_plans/experiment_registry.py
@@ -6,6 +6,7 @@ from dodal.devices.fast_grid_scan import GridScanParams
 from dodal.devices.rotation_scan import RotationScanParams
 
 from artemis.experiment_plans import fast_grid_scan_plan
+from artemis.parameters.plan_specific.fgs_internal_params import FGSInternalParameters
 
 
 def not_implemented():
@@ -21,7 +22,7 @@ PLAN_REGISTRY: Dict[str, Dict[str, Callable]] = {
     "fast_grid_scan": {
         "setup": fast_grid_scan_plan.create_devices,
         "run": fast_grid_scan_plan.get_plan,
-        "internal_param_type": fast_grid_scan_plan.FGSInternalParameters,
+        "internal_param_type": FGSInternalParameters,
         "experiment_param_type": GridScanParams,
     },
     "rotation_scan": {
