+++ src/artemis/parameters/plan_specific/fgs_internal_params.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from typing import Any
+
+from dodal.devices.fast_grid_scan import GridScanParams
+
+from artemis.parameters.internal_parameters import InternalParameters
+
+
+class FGSInternalParameters(InternalParameters):
+    experiment_params_type = GridScanParams
+
+    def pre_sorting_translation(self, param_dict: dict[str, Any]):
+        super().pre_sorting_translation(param_dict)
+        param_dict["omega_increment"] = 0
+        param_dict["num_triggers"] = param_dict["num_images"]
+        param_dict["num_images_per_trigger"] = 1
